requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
difflib==3.10.0
pymongo==4.6.1